[refactor/#24] API 호출 함수 이동에 따른 경로 수정
